requirements.txt
os
google-genai
dotenv
pandas
reportlab
IPython
json
time